requirements.txt
python-telegram-bot[webhooks]
supabase
python-dotenv
uvicorn
httpx
